disease_detection/requirements.txt
fastapi
uvicorn
torch
timm
Pillow
huggingface_hub
python-multipart